+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-* Adding `ProteinEmbedding` class and corresponding file format ([#2008](https://github.com/scikit-bio/scikit-bio/pull/2008]))
+* Added `ProteinEmbedding` class and corresponding file format ([#2008](https://github.com/scikit-bio/scikit-bio/pull/2008]))
 
 ## Version 0.6.0